+++ packages/hydrogen/src/hooks/useShopQuery/hooks.ts
@@ -38,7 +38,7 @@ export function useShopQuery<T>({
   const body = query ? graphqlRequestBody(query, variables) : '';
   const {request, key} = createShopRequest(body);
 
-  const {data} = useQuery<UseShopQueryResponse<T>>(
+  const {data, error: fetchError} = useQuery<UseShopQueryResponse<T>>(
     key,
     query
       ? fetchBuilder<UseShopQueryResponse<T>>(request)
@@ -46,6 +46,29 @@ export function useShopQuery<T>({
         async () => ({data: undefined as unknown as T, errors: undefined}),
     {cache}
   );
+
+  /**
+   * The fetch request itself failed, so we handle that differently than a GraphQL error
+   */
+  if (fetchError) {
+    const errorMessage = `Failed to fetch the Storefront API. ${
+      // 403s to the SF API (almost?) always mean that your Shopify credentials are bad/wrong
+      fetchError.status === 403
+        ? `You may have a bad value in 'shopify.config.js'`
+        : `${fetchError.statusText}`
+    }`;
+
+    log.error(errorMessage);
+
+    if (getConfig().dev) {
+      throw new Error(errorMessage);
+    } else {
+      // in non-dev environments, we probably don't want super-detailed error messages for the user
+      throw new Error(
+        `The fetch attempt failed; there was an issue connecting to the data source.`
+      );
+    }
+  }
 
   /**
    * GraphQL errors get printed to the console but ultimately