+++ packages/hydrogen/src/utilities/fetch.ts
@@ -46,6 +46,11 @@ export function fetchBuilder<T>(request: Request) {
       headers,
       method: clonedRequest.method,
     });
+
+    if (!response.ok) {
+      throw response;
+    }
+
     const data = await response.json();
 
     return data as T;